first_service/requirements.txt
grpcio-tools==1.43.0
grpcio==1.43.0
grpc-interceptor==0.12.0
pytest==5.4
